Fixes the compilation error on Solaris due to a missing LOG_PERROR.
The LOG_PERROR symbol is not part of the POSIX specification, though it
is defined on most Unix variants. This caused compilation to fail on
Solaris systems, since the symbol is not defined.

Using LOG_PERROR functionality on one of these sytems will now surface a
warning to the user instead.

Signed-off-by: Aaron Jacobs <aaron.jacobs@crescendotechnology.com>